chore: align README and move default facet-upgrade into config subdir
