+++ scripts/upgrade-facets.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 const hre = require("hardhat");
 const { ZeroAddress, getContractAt, provider, getSigners, getContractFactory } = hre.ethers;
 const network = hre.network.name;
-const { getFacets } = require("./config/facet-upgrade");
+const { getFacets } = require("./config/upgrade/facet-upgrade");
 const environments = require("../environments");
 const tipMultiplier = BigInt(environments.tipMultiplier);
 const tipSuggestion = 1500000000n; // js always returns this constant, it does not vary per block
